refactored ingredeint quantity validation and refactored quantity column in ingredient_recipe migration

+++ database/migrations/2025_02_17_000012_ingredient_recipe.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('recipe_id')->constrained()->cascadeOnDelete();
             $table->foreignId('ingredient_id')->constrained()->cascadeOnDelete();
-            $table->integer('quantity');
+            $table->decimal('quantity', 8, 2);
             $table->foreignId('unit_id')->constrained();
             $table->timestamps();
         });
